DAOS-6083 control: Improve server-side RPC handling
Mostly a code re-org, to better reflect the different MS
responsibilities and break up large files into smaller ones.

Adds new helpers to sanity-check that a MS request has been
directed to a valid replica or the active leader, depending on
the RPC.

Removes the panic() calls in the raft FSM in favor of an
emergency shutdown facility that noisily crashes the local node
out of the raft cluster but leaves the control plane server running.

Also removes some dead code associated with the old MS.
