+++ src/control/common/types.go
@@ -23,7 +23,10 @@
 
 package common
 
-import "fmt"
+import (
+	"fmt"
+	"reflect"
+)
 
 // AccessControlList is a structure for the access control list.
 type AccessControlList struct {
@@ -76,4 +79,21 @@ func (acl *AccessControlList) String() string {
 type PoolDiscovery struct {
 	UUID        string   // Unique identifier
 	SvcReplicas []uint32 `json:"Svcreps"` // Ranks of pool service replicas
+}
+
+// InterfaceIsNil returns true if the interface itself or its underlying value
+// is nil.
+func InterfaceIsNil(i interface{}) bool {
+	// NB: The unsafe version of this, which makes assumptions
+	// about the memory layout of interfaces, is a constant
+	// 2.54 ns/op.
+	// return (*[2]uintptr)(unsafe.Pointer(&i))[1] == 0
+
+	// This version is only slightly slower in the grand scheme
+	// of things. 3.96 ns/op for a nil type/value, and 7.98 ns/op
+	// for a non-nil value.
+	if i == nil {
+		return true
+	}
+	return reflect.ValueOf(i).IsNil()
 }
