+++ src/control/server/instance_superblock_test.go
@@ -24,7 +24,6 @@
 package server
 
 import (
-	"net"
 	"os"
 	"path/filepath"
 	"testing"
@@ -42,7 +41,6 @@ func TestServer_Instance_createSuperblock(t *testing.T) {
 	testDir, cleanup := CreateTestDir(t)
 	defer cleanup()
 
-	ctrlAddrs := []string{"1.2.3.4:5", "6.7.8.9:10"}
 	h := NewIOServerHarness(log)
 	for idx, mnt := range []string{"one", "two"} {
 		if err := os.MkdirAll(filepath.Join(testDir, mnt), 0777); err != nil {
@@ -64,23 +62,6 @@ func TestServer_Instance_createSuperblock(t *testing.T) {
 		if err := h.AddInstance(srv); err != nil {
 			t.Fatal(err)
 		}
-	}
-
-	// ugh, this isn't ideal
-	oldGetAddrFn := getInterfaceAddrs
-	defer func() {
-		getInterfaceAddrs = oldGetAddrFn
-	}()
-	getInterfaceAddrs = func() ([]net.Addr, error) {
-		addrs := make([]net.Addr, len(ctrlAddrs))
-		var err error
-		for i, ca := range ctrlAddrs {
-			addrs[i], err = net.ResolveTCPAddr("tcp", ca)
-			if err != nil {
-				return nil, err
-			}
-		}
-		return addrs, nil
 	}
 
 	for _, instance := range h.Instances() {
