+++ src/control/server/mgmt_svc.go
@@ -24,17 +24,16 @@
 package server
 
 import (
-	"fmt"
 	"net"
-	"os/exec"
-	"strconv"
+	"time"
 
 	"github.com/golang/protobuf/proto"
+	uuid "github.com/google/uuid"
 	"github.com/pkg/errors"
 	"golang.org/x/net/context"
+	"google.golang.org/grpc/peer"
 
 	"github.com/daos-stack/daos/src/control/common"
-	"github.com/daos-stack/daos/src/control/common/proto/convert"
 	mgmtpb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
 	"github.com/daos-stack/daos/src/control/drpc"
 	"github.com/daos-stack/daos/src/control/logging"
@@ -42,104 +41,25 @@ import (
 	"github.com/daos-stack/daos/src/control/system"
 )
 
-// CheckReplica verifies if this server is supposed to host an MS replica,
-// only performing the check and printing the result for now.
-func CheckReplica(lis net.Listener, accessPoints []string, srv *exec.Cmd) (msReplicaCheck string, err error) {
-	isReplica, bootstrap, err := checkMgmtSvcReplica(lis.Addr().(*net.TCPAddr), accessPoints)
-	if err != nil {
-		_ = srv.Process.Kill()
-		return
+const (
+	batchJoinInterval = 250 * time.Millisecond
+	joinRespTimeout   = 10 * time.Millisecond
+)
+
+type (
+	batchJoinRequest struct {
+		mgmtpb.JoinReq
+		peerAddr *net.TCPAddr
+		respCh   chan *batchJoinResponse
 	}
 
-	if isReplica {
-		msReplicaCheck = " as access point"
-		if bootstrap {
-			msReplicaCheck += " (bootstrap)"
-		}
+	batchJoinResponse struct {
+		mgmtpb.JoinResp
+		joinErr error
 	}
 
-	return
-}
-
-// getInterfaceAddrs enables TestCheckMgmtSvcReplica to replace the real
-// interface query with a sample data set.
-var getInterfaceAddrs = func() ([]net.Addr, error) {
-	return net.InterfaceAddrs()
-}
-
-// checkMgmtSvcReplica determines if this server is supposed to host an MS
-// replica, based on this server's management address and the system access
-// points. If bootstrap is true, in which case isReplica must be true, this
-// replica shall bootstrap the MS.
-func checkMgmtSvcReplica(self *net.TCPAddr, accessPoints []string) (isReplica, bootstrap bool, err error) {
-	replicas, err := resolveAccessPoints(accessPoints)
-	if err != nil {
-		return false, false, err
-	}
-
-	selves, err := getListenIPs(self)
-	if err != nil {
-		return false, false, err
-	}
-
-	// Check each replica against this server's listen IPs.
-	for i := range replicas {
-		if replicas[i].Port != self.Port {
-			continue
-		}
-		for _, ip := range selves {
-			if ip.Equal(replicas[i].IP) {
-				// The first replica in the access point list
-				// shall bootstrap the MS.
-				if i == 0 {
-					return true, true, nil
-				}
-				return true, false, nil
-			}
-		}
-	}
-
-	return false, false, nil
-}
-
-// resolveAccessPoints resolves the strings in accessPoints into addresses in
-// addrs. If a port isn't specified, assume the default port.
-func resolveAccessPoints(accessPoints []string) (addrs []*net.TCPAddr, err error) {
-	defaultPort := config.DefaultServer().ControlPort
-	for _, ap := range accessPoints {
-		if !common.HasPort(ap) {
-			ap = net.JoinHostPort(ap, strconv.Itoa(defaultPort))
-		}
-		t, err := net.ResolveTCPAddr("tcp", ap)
-		if err != nil {
-			return nil, err
-		}
-		addrs = append(addrs, t)
-	}
-	return addrs, nil
-}
-
-// getListenIPs takes the address this server listens on and returns a list of
-// the corresponding IPs.
-func getListenIPs(listenAddr *net.TCPAddr) (listenIPs []net.IP, err error) {
-	if listenAddr.IP.IsUnspecified() {
-		// Find the IPs of all IP interfaces.
-		addrs, err := getInterfaceAddrs()
-		if err != nil {
-			return nil, err
-		}
-		for _, a := range addrs {
-			// Ignore non-IP interfaces.
-			in, ok := a.(*net.IPNet)
-			if ok {
-				listenIPs = append(listenIPs, in.IP)
-			}
-		}
-	} else {
-		listenIPs = append(listenIPs, listenAddr.IP)
-	}
-	return listenIPs, nil
-}
+	joinReqChan chan *batchJoinRequest
+)
 
 // mgmtSvc implements (the Go portion of) Management Service, satisfying
 // mgmtpb.MgmtSvcServer.
@@ -163,7 +83,39 @@ func newMgmtSvc(h *IOServerHarness, m *system.Membership, s *system.Database) *m
 	}
 }
 
+// checkLeaderRequest performs sanity-checking on a request that must
+// be run on the current MS leader.
+func (svc *mgmtSvc) checkLeaderRequest(req proto.Message) error {
+	if err := svc.sysdb.CheckLeader(); err != nil {
+		return err
+	}
+	if common.InterfaceIsNil(req) {
+		return errors.New("nil request")
+	}
+	return nil
+}
+
+// checkReplicaRequest performs sanity-checking on a request that must
+// be run on a MS replica.
+func (svc *mgmtSvc) checkReplicaRequest(req proto.Message) error {
+	if err := svc.sysdb.CheckReplica(); err != nil {
+		return err
+	}
+	if common.InterfaceIsNil(req) {
+		return errors.New("nil request")
+	}
+	return nil
+}
+
+// GetAttachInfo handles a request to retrieve a map of ranks to fabric URIs, in addition
+// to client network autoconfiguration hints.
+//
+// The default use case is for libdaos clients to obtain the set of ranks associated
+// with MS replicas.
 func (svc *mgmtSvc) GetAttachInfo(ctx context.Context, req *mgmtpb.GetAttachInfoReq) (*mgmtpb.GetAttachInfoResp, error) {
+	if err := svc.checkReplicaRequest(req); err != nil {
+		return nil, err
+	}
 	svc.log.Debugf("MgmtSvc.GetAttachInfo dispatch, req:%+v\n", *req)
 
 	if svc.clientNetworkCfg == nil {
@@ -197,354 +149,282 @@ func (svc *mgmtSvc) GetAttachInfo(ctx context.Context, req *mgmtpb.GetAttachInfo
 	return resp, nil
 }
 
-func queryRank(reqRank uint32, srvRank system.Rank) bool {
-	rr := system.Rank(reqRank)
-	if rr.Equals(system.NilRank) {
-		return true
+// LeaderQuery returns the system leader and access point replica details.
+func (svc *mgmtSvc) LeaderQuery(ctx context.Context, req *mgmtpb.LeaderQueryReq) (*mgmtpb.LeaderQueryResp, error) {
+	if req == nil {
+		return nil, errors.New("nil request")
 	}
-	return rr.Equals(srvRank)
+
+	svc.log.Debugf("MgmtSvc.LeaderQuery dispatch, req:%+v\n", req)
+
+	if req.System != svc.sysdb.SystemName() {
+		return nil, errors.Errorf("received leader query for wrong system (local: %q, req: %q)",
+			svc.sysdb.SystemName(), req.System)
+	}
+
+	leaderAddr, replicas, err := svc.sysdb.LeaderQuery()
+	if err != nil {
+		return nil, err
+	}
+
+	resp := &mgmtpb.LeaderQueryResp{
+		CurrentLeader: leaderAddr,
+		Replicas:      replicas,
+	}
+
+	svc.log.Debugf("MgmtSvc.LeaderQuery dispatch, resp:%+v\n", resp)
+	return resp, nil
 }
 
-func (svc *mgmtSvc) querySmdDevices(ctx context.Context, req *mgmtpb.SmdQueryReq, resp *mgmtpb.SmdQueryResp) error {
-	for _, srv := range svc.harness.Instances() {
-		if !srv.isReady() {
-			svc.log.Debugf("skipping not-ready instance")
-			continue
-		}
+// getPeerListenAddr combines peer ip from supplied context with input port.
+func getPeerListenAddr(ctx context.Context, listenAddrStr string) (*net.TCPAddr, error) {
+	p, ok := peer.FromContext(ctx)
+	if !ok {
+		return nil, errors.New("peer details not found in context")
+	}
 
-		srvRank, err := srv.GetRank()
-		if err != nil {
-			return err
-		}
-		if !queryRank(req.GetRank(), srvRank) {
-			continue
-		}
+	tcpAddr, ok := p.Addr.(*net.TCPAddr)
+	if !ok {
+		return nil, errors.Errorf("peer address (%s) not tcp", p.Addr)
+	}
 
-		rResp := new(mgmtpb.SmdQueryResp_RankResp)
-		rResp.Rank = srvRank.Uint32()
+	// what port is the input address listening on?
+	_, portStr, err := net.SplitHostPort(listenAddrStr)
+	if err != nil {
+		return nil, errors.Wrap(err, "get listening port")
+	}
 
-		listDevsResp, err := srv.listSmdDevices(ctx, new(mgmtpb.SmdDevReq))
-		if err != nil {
-			return errors.Wrapf(err, "rank %d", srvRank)
-		}
+	// resolve combined IP/port address
+	return net.ResolveTCPAddr(p.Addr.Network(),
+		net.JoinHostPort(tcpAddr.IP.String(), portStr))
+}
 
-		if err := convert.Types(listDevsResp.Devices, &rResp.Devices); err != nil {
-			return errors.Wrap(err, "failed to convert device list")
-		}
-		resp.Ranks = append(resp.Ranks, rResp)
+func (svc *mgmtSvc) startJoinLoop(ctx context.Context) {
+	svc.log.Debug("starting joinLoop")
+	go svc.joinLoop(ctx)
+}
 
-		if req.Uuid != "" {
-			found := false
-			for _, dev := range rResp.Devices {
-				if dev.Uuid == req.Uuid {
-					rResp.Devices = []*mgmtpb.SmdQueryResp_Device{dev}
-					found = true
-					break
-				}
-			}
-			if !found {
-				rResp.Devices = nil
-			}
-		}
+func (svc *mgmtSvc) joinLoop(parent context.Context) {
+	var joinReqs []*batchJoinRequest
 
-		if req.Target != "" {
-			reqTgtId, err := strconv.ParseInt(req.Target, 10, 32)
-			if err != nil {
-				return errors.Errorf("invalid target idx %q", req.Target)
+	for {
+		select {
+		case <-parent.Done():
+			svc.log.Debug("stopped joinLoop")
+			return
+		case jr := <-svc.joinReqs:
+			joinReqs = append(joinReqs, jr)
+		case <-time.After(batchJoinInterval):
+			if len(joinReqs) == 0 {
+				continue
 			}
 
-			found := false
-			for _, dev := range rResp.Devices {
-				for _, tgtId := range dev.TgtIds {
-					if int32(reqTgtId) == tgtId {
-						rResp.Devices = []*mgmtpb.SmdQueryResp_Device{dev}
-						found = true
-						break
+			svc.log.Debugf("processing %d join requests", len(joinReqs))
+			joinResps := make([]*batchJoinResponse, len(joinReqs))
+			for i, req := range joinReqs {
+				joinResps[i] = svc.join(parent, req)
+			}
+
+			for i := 0; i < len(svc.harness.Instances()); i++ {
+				if err := svc.doGroupUpdate(parent); err != nil {
+					if err == instanceNotReady {
+						svc.log.Debug("group update not ready (retrying)")
+						continue
+					}
+
+					err = errors.Wrap(err, "failed to perform CaRT group update")
+					for i, jr := range joinResps {
+						if jr.joinErr == nil {
+							joinResps[i] = &batchJoinResponse{joinErr: err}
+						}
 					}
 				}
+				break
 			}
-			if !found {
-				rResp.Devices = nil
-			}
-		}
 
-		if !req.IncludeBioHealth {
-			continue
-		}
+			svc.log.Debugf("sending %d join responses", len(joinReqs))
+			for i, req := range joinReqs {
+				ctx, cancel := context.WithTimeout(parent, joinRespTimeout)
+				defer cancel()
 
-		for _, dev := range rResp.Devices {
-			health, err := srv.getBioHealth(ctx, &mgmtpb.BioHealthReq{
-				DevUuid: dev.Uuid,
-			})
-			if err != nil {
-				return errors.Wrapf(err, "device %s", dev)
-			}
-			dev.Health = health
-		}
-	}
-	return nil
-}
-
-func (svc *mgmtSvc) querySmdPools(ctx context.Context, req *mgmtpb.SmdQueryReq, resp *mgmtpb.SmdQueryResp) error {
-	for _, srv := range svc.harness.Instances() {
-		if !srv.isReady() {
-			svc.log.Debugf("skipping not-ready instance")
-			continue
-		}
-
-		srvRank, err := srv.GetRank()
-		if err != nil {
-			return err
-		}
-		if !queryRank(req.GetRank(), srvRank) {
-			continue
-		}
-
-		rResp := new(mgmtpb.SmdQueryResp_RankResp)
-		rResp.Rank = srvRank.Uint32()
-
-		dresp, err := srv.CallDrpc(ctx, drpc.MethodSmdPools, new(mgmtpb.SmdPoolReq))
-		if err != nil {
-			return err
-		}
-
-		rankDevResp := new(mgmtpb.SmdPoolResp)
-		if err = proto.Unmarshal(dresp.Body, rankDevResp); err != nil {
-			return errors.Wrap(err, "unmarshal SmdListPools response")
-		}
-
-		if rankDevResp.Status != 0 {
-			return errors.Wrapf(drpc.DaosStatus(rankDevResp.Status), "rank %d ListPools failed", srvRank)
-		}
-
-		if err := convert.Types(rankDevResp.Pools, &rResp.Pools); err != nil {
-			return errors.Wrap(err, "failed to convert pool list")
-		}
-		resp.Ranks = append(resp.Ranks, rResp)
-
-		if req.Uuid != "" {
-			found := false
-			for _, pool := range rResp.Pools {
-				if pool.Uuid == req.Uuid {
-					rResp.Pools = []*mgmtpb.SmdQueryResp_Pool{pool}
-					found = true
-					break
+				select {
+				case <-ctx.Done():
+					svc.log.Errorf("failed to send join response: %s", ctx.Err())
+				case req.respCh <- joinResps[i]:
 				}
 			}
-			if !found {
-				rResp.Pools = nil
-			}
+
+			joinReqs = nil
+		}
+	}
+}
+
+func (svc *mgmtSvc) join(ctx context.Context, req *batchJoinRequest) *batchJoinResponse {
+	uuid, err := uuid.Parse(req.GetUuid())
+	if err != nil {
+		return &batchJoinResponse{
+			joinErr: errors.Wrapf(err, "invalid uuid %q", req.GetUuid()),
 		}
 	}
 
+	fd, err := system.NewFaultDomainFromString(req.GetSrvFaultDomain())
+	if err != nil {
+		return &batchJoinResponse{
+			joinErr: errors.Wrapf(err, "invalid server fault domain %q", req.GetSrvFaultDomain()),
+		}
+	}
+
+	joinResponse, err := svc.membership.Join(&system.JoinRequest{
+		Rank:           system.Rank(req.Rank),
+		UUID:           uuid,
+		ControlAddr:    req.peerAddr,
+		FabricURI:      req.GetUri(),
+		FabricContexts: req.GetNctxs(),
+		FaultDomain:    fd,
+	})
+	if err != nil {
+		return &batchJoinResponse{joinErr: err}
+	}
+
+	member := joinResponse.Member
+	if joinResponse.Created {
+		svc.log.Debugf("new system member: rank %d, addr %s, uri %s",
+			member.Rank, req.peerAddr, member.FabricURI)
+	} else {
+		svc.log.Debugf("updated system member: rank %d, uri %s, %s->%s",
+			member.Rank, member.FabricURI, joinResponse.PrevState, member.State())
+	}
+
+	resp := &batchJoinResponse{
+		JoinResp: mgmtpb.JoinResp{
+			State: mgmtpb.JoinResp_IN,
+			Rank:  member.Rank.Uint32(),
+		},
+	}
+
+	// If the rank is local to the MS leader, then we need to wire up at least
+	// one in order to perform a CaRT group update.
+	if common.IsLocalAddr(req.peerAddr) && req.Idx == 0 {
+		resp.LocalJoin = true
+
+		srvs := svc.harness.Instances()
+		if len(srvs) == 0 {
+			return &batchJoinResponse{
+				joinErr: errors.New("invalid Join request (index 0 doesn't exist?!?)"),
+			}
+		}
+		srv := srvs[0]
+
+		if err := srv.callSetRank(ctx, joinResponse.Member.Rank); err != nil {
+			return &batchJoinResponse{
+				joinErr: errors.Wrap(err, "failed to set rank on local instance"),
+			}
+		}
+
+		if err := srv.callSetUp(ctx); err != nil {
+			return &batchJoinResponse{
+				joinErr: errors.Wrap(err, "failed to load local instance modules"),
+			}
+		}
+
+		// mark the ioserver as ready to handle dRPC requests
+		srv.ready.SetTrue()
+	}
+
+	return resp
+}
+
+func (svc *mgmtSvc) doGroupUpdate(ctx context.Context) error {
+	gm, err := svc.sysdb.GroupMap()
+	if err != nil {
+		return err
+	}
+	if len(gm.RankURIs) == 0 {
+		return system.ErrEmptyGroupMap
+	}
+
+	req := &mgmtpb.GroupUpdateReq{
+		MapVersion: gm.Version,
+	}
+	rankSet := &system.RankSet{}
+	for rank, uri := range gm.RankURIs {
+		req.Servers = append(req.Servers, &mgmtpb.GroupUpdateReq_Server{
+			Rank: rank.Uint32(),
+			Uri:  uri,
+		})
+		rankSet.Add(rank)
+	}
+
+	svc.log.Debugf("group update request: version: %d, ranks: %s", req.MapVersion, rankSet)
+	dResp, err := svc.harness.CallDrpc(ctx, drpc.MethodGroupUpdate, req)
+	if err != nil {
+		if err == instanceNotReady {
+			return err
+		}
+		svc.log.Errorf("dRPC GroupUpdate call failed: %s", err)
+		return err
+	}
+
+	resp := new(mgmtpb.GroupUpdateResp)
+	if err = proto.Unmarshal(dResp.Body, resp); err != nil {
+		return errors.Wrap(err, "unmarshal GroupUpdate response")
+	}
+
+	if resp.GetStatus() != 0 {
+		return drpc.DaosStatus(resp.GetStatus())
+	}
 	return nil
 }
 
-func (svc *mgmtSvc) smdQueryDevice(ctx context.Context, req *mgmtpb.SmdQueryReq) (system.Rank, *mgmtpb.SmdQueryResp_Device, error) {
-	rank := system.NilRank
-	if req.Uuid == "" {
-		return rank, nil, errors.New("empty UUID in device query")
+// Join management service gRPC handler receives Join requests from
+// control-plane instances attempting to register a managed instance (will be a
+// rank once joined) to the DAOS system.
+//
+// On receipt of the join request, add to a queue of requests to be processed
+// periodically in a dedicated goroutine. This architecture provides for thread
+// safety and improved performance while updating the system membership and CaRT
+// primary group in the local ioserver.
+//
+// The state of the newly joined/evicted rank along with the reply address used
+// to contact the new rank in future will be registered in the system membership.
+// The reply address is generated by combining peer (sender) IP (from context)
+// with listening port from joining instance's host addr contained in the
+// provided request.
+func (svc *mgmtSvc) Join(ctx context.Context, req *mgmtpb.JoinReq) (*mgmtpb.JoinResp, error) {
+	if err := svc.checkLeaderRequest(req); err != nil {
+		return nil, err
+	}
+	svc.log.Debugf("MgmtSvc.Join dispatch, req:%#v\n", req)
+
+	replyAddr, err := getPeerListenAddr(ctx, req.GetAddr())
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to parse %q into a peer control address", req.GetAddr())
 	}
 
-	resp := new(mgmtpb.SmdQueryResp)
-	if err := svc.querySmdDevices(ctx, req, resp); err != nil {
-		return rank, nil, err
+	bjr := &batchJoinRequest{
+		JoinReq:  *req,
+		peerAddr: replyAddr,
+		respCh:   make(chan *batchJoinResponse),
 	}
 
-	for _, rr := range resp.Ranks {
-		switch len(rr.Devices) {
-		case 0:
-			continue
-		case 1:
-			rank = system.Rank(rr.Rank)
-			return rank, rr.Devices[0], nil
-		default:
-			return rank, nil, errors.Errorf("device query on %s matched multiple devices", req.Uuid)
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case svc.joinReqs <- bjr:
+	}
+
+	var resp *mgmtpb.JoinResp
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case r := <-bjr.respCh:
+		if r.joinErr != nil {
+			return nil, r.joinErr
 		}
+		resp = &r.JoinResp
 	}
 
-	return rank, nil, nil
-}
-
-func (svc *mgmtSvc) smdSetFaulty(ctx context.Context, req *mgmtpb.SmdQueryReq) (*mgmtpb.SmdQueryResp, error) {
-	req.Rank = uint32(system.NilRank)
-	rank, device, err := svc.smdQueryDevice(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	if device == nil {
-		return nil, errors.Errorf("smdSetFaulty on %s did not match any devices", req.Uuid)
-	}
-
-	srvs, err := svc.harness.FilterInstancesByRankSet(fmt.Sprintf("%d", rank))
-	if err != nil {
-		return nil, err
-	}
-	if len(srvs) == 0 {
-		return nil, errors.Errorf("failed to retrieve instance for rank %d", rank)
-	}
-
-	svc.log.Debugf("calling set-faulty on rank %d for %s", rank, req.Uuid)
-
-	dresp, err := srvs[0].CallDrpc(ctx, drpc.MethodSetFaultyState, &mgmtpb.DevStateReq{
-		DevUuid: req.Uuid,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	dsr := &mgmtpb.DevStateResp{}
-	if err = proto.Unmarshal(dresp.Body, dsr); err != nil {
-		return nil, errors.Wrap(err, "unmarshal StorageSetFaulty response")
-	}
-
-	if dsr.Status != 0 {
-		return nil, errors.Wrap(drpc.DaosStatus(dsr.Status), "smdSetFaulty failed")
-	}
-
-	return &mgmtpb.SmdQueryResp{
-		Ranks: []*mgmtpb.SmdQueryResp_RankResp{
-			{
-				Rank: rank.Uint32(),
-				Devices: []*mgmtpb.SmdQueryResp_Device{
-					{
-						Uuid:  dsr.DevUuid,
-						State: dsr.DevState,
-					},
-				},
-			},
-		},
-	}, nil
-}
-
-func (svc *mgmtSvc) smdReplace(ctx context.Context, req *mgmtpb.SmdQueryReq) (*mgmtpb.SmdQueryResp, error) {
-	req.Rank = uint32(system.NilRank)
-	rank, device, err := svc.smdQueryDevice(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	if device == nil {
-		return nil, errors.Errorf("smdReplace on %s did not match any devices", req.Uuid)
-	}
-
-	srvs, err := svc.harness.FilterInstancesByRankSet(fmt.Sprintf("%d", rank))
-	if err != nil {
-		return nil, err
-	}
-	if len(srvs) == 0 {
-		return nil, errors.Errorf("failed to retrieve instance for rank %d", rank)
-	}
-
-	svc.log.Debugf("calling storage replace on rank %d for %s", rank, req.Uuid)
-
-	dresp, err := srvs[0].CallDrpc(ctx, drpc.MethodReplaceStorage, &mgmtpb.DevReplaceReq{
-		OldDevUuid: req.Uuid,
-		NewDevUuid: req.ReplaceUUID,
-		NoReint:    req.NoReint,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	drr := &mgmtpb.DevReplaceResp{}
-	if err = proto.Unmarshal(dresp.Body, drr); err != nil {
-		return nil, errors.Wrap(err, "unmarshal StorageReplace response")
-	}
-
-	if drr.Status != 0 {
-		return nil, errors.Wrap(drpc.DaosStatus(drr.Status), "smdReplace failed")
-	}
-
-	return &mgmtpb.SmdQueryResp{
-		Ranks: []*mgmtpb.SmdQueryResp_RankResp{
-			{
-				Rank: rank.Uint32(),
-				Devices: []*mgmtpb.SmdQueryResp_Device{
-					{
-						Uuid:  drr.NewDevUuid,
-						State: drr.DevState,
-					},
-				},
-			},
-		},
-	}, nil
-}
-
-func (svc *mgmtSvc) SmdQuery(ctx context.Context, req *mgmtpb.SmdQueryReq) (*mgmtpb.SmdQueryResp, error) {
-	svc.log.Debugf("MgmtSvc.SmdQuery dispatch, req:%+v\n", *req)
-
-	if !svc.harness.isStarted() {
-		return nil, FaultHarnessNotStarted
-	}
-	if len(svc.harness.readyRanks()) == 0 {
-		return nil, FaultDataPlaneNotStarted
-	}
-
-	if req.SetFaulty {
-		return svc.smdSetFaulty(ctx, req)
-	}
-
-	if req.ReplaceUUID != "" {
-		return svc.smdReplace(ctx, req)
-	}
-
-	if req.Uuid != "" && (!req.OmitDevices && !req.OmitPools) {
-		return nil, errors.New("UUID is ambiguous when querying both pools and devices")
-	}
-	if req.Target != "" && req.Rank == uint32(system.NilRank) {
-		return nil, errors.New("Target is invalid without Rank")
-	}
-
-	resp := new(mgmtpb.SmdQueryResp)
-	if !req.OmitDevices {
-		if err := svc.querySmdDevices(ctx, req, resp); err != nil {
-			return nil, err
-		}
-	}
-	if !req.OmitPools {
-		if err := svc.querySmdPools(ctx, req, resp); err != nil {
-			return nil, err
-		}
-	}
-
-	svc.log.Debugf("MgmtSvc.SmdQuery dispatch, resp:%+v\n", *resp)
-	return resp, nil
-}
-
-// ListContainers implements the method defined for the Management Service.
-func (svc *mgmtSvc) ListContainers(ctx context.Context, req *mgmtpb.ListContReq) (*mgmtpb.ListContResp, error) {
-	svc.log.Debugf("MgmtSvc.ListContainers dispatch, req:%+v\n", *req)
-
-	dresp, err := svc.harness.CallDrpc(ctx, drpc.MethodListContainers, req)
-	if err != nil {
-		return nil, err
-	}
-
-	resp := &mgmtpb.ListContResp{}
-	if err = proto.Unmarshal(dresp.Body, resp); err != nil {
-		return nil, errors.Wrap(err, "unmarshal ListContainers response")
-	}
-
-	svc.log.Debugf("MgmtSvc.ListContainers dispatch, resp:%+v\n", *resp)
-
-	return resp, nil
-}
-
-// ContSetOwner forwards a gRPC request to the DAOS IO server to change a container's ownership.
-func (svc *mgmtSvc) ContSetOwner(ctx context.Context, req *mgmtpb.ContSetOwnerReq) (*mgmtpb.ContSetOwnerResp, error) {
-	svc.log.Debugf("MgmtSvc.ContSetOwner dispatch, req:%+v\n", *req)
-
-	dresp, err := svc.harness.CallDrpc(ctx, drpc.MethodContSetOwner, req)
-	if err != nil {
-		return nil, err
-	}
-
-	resp := &mgmtpb.ContSetOwnerResp{}
-	if err = proto.Unmarshal(dresp.Body, resp); err != nil {
-		return nil, errors.Wrap(err, "unmarshal ContSetOwner response")
-	}
-
-	svc.log.Debugf("MgmtSvc.ContSetOwner dispatch, resp:%+v\n", *resp)
+	svc.log.Debugf("MgmtSvc.Join dispatch, resp:%#v\n", resp)
 
 	return resp, nil
 }
